Rename utils.py->uei.py, update serializer import
