+++ backend/api/serializers.py
@@ -4,7 +4,7 @@ from rest_framework import serializers
 from django.utils.translation import gettext_lazy as _
 
 from audit.models import SingleAuditChecklist, Access
-from api.utils import get_uei_info_from_sam_gov
+from api.uei import get_uei_info_from_sam_gov
 
 
 SPENDING_THRESHOLD = _("The FAC only accepts submissions from non-Federal entities that spend $750,000 or more in federal awards during its audit period (fiscal period begin dates on or after 12/26/2014) in accordance with Uniform Guidance")
